+++ cryptolyzer/common/dhparam.py
@@ -14,10 +14,18 @@ from .math import is_prime, prime_precheck
 
 @attr.s(eq=False)
 class DHParameterNumbers(object):
-    p = attr.ib(validator=attr.validators.instance_of(six.integer_types))  # pylint: disable=invalid-name
-    g = attr.ib(validator=attr.validators.instance_of(six.integer_types))  # pylint: disable=invalid-name
+    p = attr.ib(  # pylint: disable=invalid-name
+        validator=attr.validators.instance_of(six.integer_types),
+        metadata={'human_readable_name': 'p'},
+    )
+    g = attr.ib(  # pylint: disable=invalid-name
+        validator=attr.validators.instance_of(six.integer_types),
+        metadata={'human_readable_name': 'g'},
+    )
     q = attr.ib(  # pylint: disable=invalid-name
-        default=None, validator=attr.validators.optional(attr.validators.instance_of(six.integer_types))
+        default=None,
+        validator=attr.validators.optional(attr.validators.instance_of(six.integer_types)),
+        metadata={'human_readable_name': 'q'},
     )
 
     def __eq__(self, other):
@@ -26,7 +34,10 @@ class DHParameterNumbers(object):
 
 @attr.s
 class DHPublicNumbers(object):
-    y = attr.ib(validator=attr.validators.instance_of(six.integer_types))  # pylint: disable=invalid-name
+    y = attr.ib(  # pylint: disable=invalid-name
+        validator=attr.validators.instance_of(six.integer_types),
+        metadata={'human_readable_name': 'y'},
+    )
     parameter_numbers = attr.ib(validator=attr.validators.instance_of(DHParameterNumbers))
 
 
@@ -48,6 +59,36 @@ class DHParamWellKnown(object):
 
 
 class WellKnownDHParams(enum.Enum):
+    RFC2539_768_BIT_MODP_GROUP = DHParamWellKnown(
+        dh_param_numbers=DHParameterNumbers(
+            p=int((
+                'FFFFFFFF FFFFFFFF C90FDAA2 2168C234 C4C6628B 80DC1CD1' +
+                '29024E08 8A67CC74 020BBEA6 3B139B22 514A0879 8E3404DD' +
+                'EF9519B3 CD3A431B 302B0A6D F25F1437 4FE1356D 6D51C245' +
+                'E485B576 625E7EC6 F44C42E9 A63A3620 FFFFFFFF FFFFFFFF'
+            ).replace(' ', ''), 16),
+            g=0x2,
+        ),
+        key_size=768,
+        name='768-bit MODP Group',
+        source='RFC2539',
+    )
+    RFC2539_1024_BIT_MODP_GROUP = DHParamWellKnown(
+        dh_param_numbers=DHParameterNumbers(
+            p=int((
+                'FFFFFFFF FFFFFFFF C90FDAA2 2168C234 C4C6628B 80DC1CD1' +
+                '29024E08 8A67CC74 020BBEA6 3B139B22 514A0879 8E3404DD' +
+                'EF9519B3 CD3A431B 302B0A6D F25F1437 4FE1356D 6D51C245' +
+                'E485B576 625E7EC6 F44C42E9 A637ED6B 0BFF5CB6 F406B7ED' +
+                'EE386BFB 5A899FA5 AE9F2411 7C4B1FE6 49286651 ECE65381' +
+                'FFFFFFFF FFFFFFFF'
+            ).replace(' ', ''), 16),
+            g=0x2,
+        ),
+        key_size=1024,
+        name='1024-bit MODP Group',
+        source='RFC2539',
+    )
     RFC3526_1536_BIT_MODP_GROUP = DHParamWellKnown(
         dh_param_numbers=DHParameterNumbers(
             p=int((
@@ -652,21 +693,25 @@ class TlsDHParamVector(Vector):  # pylint: disable=too-many-ancestors
         return VectorParamNumeric(item_size=1, min_byte_num=1, max_byte_num=2 ** 16 - 1)
 
 
-def parse_dh_params(param_bytes):
+def get_dh_public_key_from_bytes(p_bytes, g_bytes, y_bytes):
+    p = int(''.join(map('{:02x}'.format, p_bytes)), 16)  # pylint: disable=invalid-name
+    g = int(''.join(map('{:02x}'.format, g_bytes)), 16)  # pylint: disable=invalid-name
+    y = int(''.join(map('{:02x}'.format, y_bytes)), 16)  # pylint: disable=invalid-name
+
+    parameter_numbers = DHParameterNumbers(p, g)
+    public_numbers = DHPublicNumbers(y, parameter_numbers)
+
+    return DHPublicKey(public_numbers, len(bytearray(p_bytes).lstrip(b'\x00')) * 8)
+
+
+def parse_tls_dh_params(param_bytes):
     parser = ParserBinary(param_bytes)
 
     parser.parse_parsable('p', TlsDHParamVector)
     parser.parse_parsable('g', TlsDHParamVector)
     parser.parse_parsable('y', TlsDHParamVector)
 
-    p = int(''.join(map('{:02x}'.format, parser['p'])), 16)  # pylint: disable=invalid-name
-    g = int(''.join(map('{:02x}'.format, parser['g'])), 16)  # pylint: disable=invalid-name
-    y = int(''.join(map('{:02x}'.format, parser['y'])), 16)  # pylint: disable=invalid-name
-
-    parameter_numbers = DHParameterNumbers(p, g)
-    public_numbers = DHPublicNumbers(y, parameter_numbers)
-
-    return DHPublicKey(public_numbers, len(parser['p']) * 8)
+    return get_dh_public_key_from_bytes(parser['p'], parser['g'], parser['y'])
 
 
 def parse_ecdh_params(param_bytes):