+++ cryptolyzer/ssh/client.py
@@ -1,17 +1,26 @@
 # -*- coding: utf-8 -*-
 
+import attr
 import six
 
+from cryptoparser.common.algorithm import KeyExchange
 from cryptoparser.common.exception import NotEnoughData
 
-from cryptoparser.ssh.record import SshRecord
-from cryptoparser.ssh.subprotocol import SshProtocolMessage, SshKeyExchangeInit, SshMessageCode
+from cryptoparser.ssh.record import SshRecordInit, SshRecordKexDH, SshRecordKexDHGroup
 from cryptoparser.ssh.ciphersuite import (
     SshCompressionAlgorithm,
     SshEncryptionAlgorithm,
     SshHostKeyAlgorithm,
     SshKexAlgorithm,
     SshMacAlgorithm,
+)
+from cryptoparser.ssh.subprotocol import (
+    SshKeyExchangeInit,
+    SshDHGroupExchangeGroup,
+    SshDHGroupExchangeInit,
+    SshDHGroupExchangeRequest,
+    SshDisconnectMessage,
+    SshProtocolMessage,
 )
 from cryptoparser.ssh.version import SshProtocolVersion, SshVersion
 
@@ -32,17 +41,46 @@ class SshProtocolMessageDefault(SshProtocolMessage):
 
 
 class SshKeyExchangeInitAnyAlgorithm(SshKeyExchangeInit):
-    def __init__(self):
+    def __init__(
+            self,
+            kex_algorithms=tuple(SshKexAlgorithm),
+            host_key_algorithms=tuple(SshHostKeyAlgorithm),
+            encryption_algorithms_client_to_server=tuple(SshEncryptionAlgorithm),
+            encryption_algorithms_server_to_client=tuple(SshEncryptionAlgorithm),
+            mac_algorithms_client_to_server=tuple(SshMacAlgorithm),
+            mac_algorithms_server_to_client=tuple(SshMacAlgorithm),
+            compression_algorithms_client_to_server=tuple(SshCompressionAlgorithm),
+            compression_algorithms_server_to_client=tuple(SshCompressionAlgorithm),
+    ):  # pylint: disable=too-many-arguments
+
         super(SshKeyExchangeInitAnyAlgorithm, self).__init__(
-            kex_algorithms=list(SshKexAlgorithm),
-            host_key_algorithms=list(SshHostKeyAlgorithm),
-            encryption_algorithms_client_to_server=list(SshEncryptionAlgorithm),
-            encryption_algorithms_server_to_client=list(SshEncryptionAlgorithm),
-            mac_algorithms_client_to_server=list(SshMacAlgorithm),
-            mac_algorithms_server_to_client=list(SshMacAlgorithm),
-            compression_algorithms_client_to_server=list(SshCompressionAlgorithm),
-            compression_algorithms_server_to_client=list(SshCompressionAlgorithm),
+            kex_algorithms=kex_algorithms,
+            host_key_algorithms=host_key_algorithms,
+            encryption_algorithms_client_to_server=encryption_algorithms_client_to_server,
+            encryption_algorithms_server_to_client=encryption_algorithms_server_to_client,
+            mac_algorithms_client_to_server=mac_algorithms_client_to_server,
+            mac_algorithms_server_to_client=mac_algorithms_server_to_client,
+            compression_algorithms_client_to_server=compression_algorithms_client_to_server,
+            compression_algorithms_server_to_client=compression_algorithms_server_to_client,
         )
+
+
+class SshKeyExchangeInitKeyExchangeDHE(SshKeyExchangeInitAnyAlgorithm):
+    def __init__(self):
+        super(SshKeyExchangeInitKeyExchangeDHE, self).__init__(
+            kex_algorithms=[
+                kex_algorithm
+                for kex_algorithm in SshKexAlgorithm
+                if kex_algorithm.value.kex == KeyExchange.DHE
+            ]
+        )
+
+
+@attr.s(frozen=True)
+class L7ServerSshGexParams(object):
+    gex_min = attr.ib(default=768, validator=attr.validators.instance_of(six.integer_types))
+    gex_max = attr.ib(default=8192, validator=attr.validators.instance_of(six.integer_types))
+    gex_number = attr.ib(default=2048, validator=attr.validators.instance_of(six.integer_types))
 
 
 class L7ClientSsh(L7TransferBase):
@@ -66,17 +104,19 @@ class L7ClientSsh(L7TransferBase):
             self,
             protocol_message=SshProtocolMessageDefault(),
             key_exchange_init_message=SshKeyExchangeInitAnyAlgorithm(),
-            last_message_type=SshKeyExchangeInit
+            gex_params=L7ServerSshGexParams(),
+            last_message_type=SshKeyExchangeInit,
     ):
         self.init_connection()
 
         try:
             ssh_client = SshClientHandshake()
             ssh_client.do_handshake(
-                self.l4_transfer,
-                protocol_message,
-                key_exchange_init_message,
-                last_message_type
+                transfer=self.l4_transfer,
+                protocol_message=protocol_message,
+                key_exchange_init_message=key_exchange_init_message,
+                gex_params=gex_params,
+                last_message_type=last_message_type,
             )
         finally:
             self._close_connection()
@@ -85,25 +125,62 @@ class L7ClientSsh(L7TransferBase):
 
 
 class SshClientHandshake(SshHandshakeBase):
+    @classmethod
+    def _process_kex_init(cls, transfer, record, key_exchange_init_message, gex_params):
+        agreed_kex = list(filter(
+            record.packet.kex_algorithms.__contains__,
+            key_exchange_init_message.kex_algorithms
+        ))
+        if not agreed_kex:
+            raise NotImplementedError()
+
+        agreed_kex_type = agreed_kex[0]
+        if agreed_kex_type.value.kex == KeyExchange.DHE:
+            if agreed_kex_type.value.key_size is None:
+                record_class = SshRecordKexDHGroup
+                transfer.send(record_class(SshDHGroupExchangeRequest(
+                    gex_min=gex_params.gex_min,
+                    gex_max=gex_params.gex_max,
+                    gex_number=gex_params.gex_number,
+                )).compose())
+                raise IndexError(record_class)
+
+            record_class = SshRecordKexDH
+        else:
+            raise NotImplementedError()
+
+        return record_class
+
+    @classmethod
+    def _process_dh_group_exchange_group(cls, record, transfer, record_class):
+        key_size = (len(record.packet.p) - 1) * 8
+        ephemeral_public_key = b'\x00' + key_size // 512 * b'\xff'
+        transfer.send(record_class(SshDHGroupExchangeInit(ephemeral_public_key)).compose())
+
+        return SshRecordKexDHGroup
+
     def do_handshake(
             self,
             transfer,
             protocol_message,
             key_exchange_init_message,
-            last_message_type
-    ):
+            gex_params,
+            last_message_type,
+    ):  # pylint: disable=too-many-arguments
         self.server_messages = self.do_key_exchange_init(
             transfer, protocol_message, key_exchange_init_message, last_message_type
         )
         if last_message_type in self.server_messages:
             return
 
+        record_class = SshRecordInit
+
         while True:
             try:
-                record, parsed_length = SshRecord.parse_immutable(transfer.buffer)
+                record, parsed_length = record_class.parse_immutable(transfer.buffer)
                 transfer.flush_buffer(parsed_length)
 
-                if record.packet.get_message_code() == SshMessageCode.DISCONNECT:
+                if isinstance(record.packet, SshDisconnectMessage):
                     raise SshDisconnect(record.packet.reason, record.packet.description)
 
                 self._last_processed_message_type = type(record.packet)
@@ -111,9 +188,17 @@ class SshClientHandshake(SshHandshakeBase):
                 if self._last_processed_message_type.get_message_code() == last_message_type:
                     break
 
-                receivable_byte_num = 0
+                if isinstance(record.packet, SshKeyExchangeInit):
+                    record_class = self._process_kex_init(transfer, record, key_exchange_init_message, gex_params)
+                elif isinstance(record.packet, SshDHGroupExchangeGroup):
+                    record_class = self._process_dh_group_exchange_group(record, transfer, record_class)
+
+                receivable_byte_num = record_class.HEADER_SIZE
             except NotEnoughData as e:
                 receivable_byte_num = e.bytes_needed
+            except IndexError as e:
+                record_class = e.args[0]
+                continue
 
             try:
                 transfer.receive(receivable_byte_num)